+++ Ejemplos_Devices/Integrada/Ejemplo_Maui_Hibrida/App.xaml.cs
@@ -3,7 +3,7 @@
 public partial class App : Application
 {
     public App() 
-    {
+    { 
         InitializeComponent();  
     }
     protected override Window CreateWindow(IActivationState? activationState)